Move YAML export functionality out of core
